♻️ Refactor Rust codebase architecture for better maintainability
Major refactoring to improve code organization and fix compilation errors:

- Split monolithic lib.rs (1471 lines) into focused modules:
  - error.rs: Centralized error handling with thiserror
  - state.rs: Consolidated application state management
  - handlers/: Menu, shortcut, and window event handlers
  - startup/: Application initialization and data loading
  - modules/: Organized command implementations

- Added proper error types (BlinkError) replacing string errors
- Fixed GlobalShortcutExt trait usage for Tauri v2 compatibility
- Fixed import paths and missing trait imports (Emitter)
- Resolved all 102 compilation errors from architecture review

The refactored code maintains all functionality while being more maintainable
and following Rust best practices for module organization.
