+++ src-tauri/Cargo.toml
@@ -35,6 +35,7 @@ sha2 = "0.10"
 tauri-plugin-clipboard-manager = "2.3.0"
 rusqlite = { version = "0.32", features = ["bundled", "chrono", "serde_json"] }
 anyhow = "1.0"
+thiserror = "1.0"
 
 [dev-dependencies]
 tempfile = "3.8"
