+++ src-tauri/src/handlers/mod.rs
@@ -1,0 +1,7 @@
+pub mod menu_handler;
+pub mod shortcut_handler;
+pub mod window_handler;
+
+pub use menu_handler::*;
+pub use shortcut_handler::*;
+pub use window_handler::*;
