+++ src-tauri/src/modules/mod.rs
@@ -2,10 +2,13 @@
 pub mod storage;
 pub mod file_storage;
 pub mod file_notes_storage;
+pub mod file_operations;
 pub mod database;
 pub mod commands;
 pub mod note_commands;
 pub mod window_commands;
+pub mod system_commands;
+pub mod test_commands;
 // pub mod config;  // TODO: Extract config functions
 pub mod windows;
 // pub mod shortcuts;  // TODO: Extract shortcuts functions
