+++ src-tauri/src/startup/mod.rs
@@ -1,0 +1,5 @@
+pub mod app_setup;
+pub mod data_loader;
+
+pub use app_setup::*;
+pub use data_loader::*;
